+++ public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/main.js": "/js/main.js",
-    "/css/app.css": "/css/app.css"
+    "/js/main.js": "/js/main.js?id=1c897fac1b7dcabcabfa",
+    "/css/app.css": "/css/app.css?id=9632034ee1f84f830e10"
 }
