+++ resources/js/components/modules/ModuleFooter.vue
