+++ resources/js/components/modules/templates/CardFlipTemplate.vue
