+++ resources/js/components/modules/templates/IntroSlide.vue
