+++ resources/js/components/modules/templates/VideoTemplate.vue
