+++ resources/js/components/modules/templates/info/InfoContent.vue
