+++ resources/js/components/modules/templates/info/InfoPhoto.vue
