+++ resources/js/components/modules/templates/quiz/QuizOption.vue
